% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I.rec.R
\name{I.rec}
\alias{I.rec}
\alias{I.rec.matrix}
\title{Indicator  of record  occurrence}
\usage{
I.rec(X_T, record = "upper")

I.rec.matrix(XM_T, record = "upper")
}
\arguments{
\item{X_T}{A numeric vector.}

\item{record}{A character string indicating the type of record  to be calculated,  "upper" or "lower".}

\item{XM_T}{A numeric matrix.}
}
\value{
A binary vector,  indicating the  record occurrence.
}
\description{
Given  a vector,  this function calculates  a binay variable  which takes the value 1 if the corresponding value in the vector is
a record and 0 otherwise.
}
\details{
Indicators of both upper and lower  records can be calculated.

The function \code{I.rec.matrix} is equivalent to \code{I.rec} but it is applied to a matrix. The  previous procedure to obtain a binary variable is applied to each
column of the matrix. In this function, argument \code{X_T} is  a matrix called \code{XM_T}, and
the return value is  a matrix  formed by the corresponding binary variables.
}
\examples{
Y1<-c(5,7,3,6,19,2,20)
I.rec(Y1)
I.rec.matrix(ZaragozaSeries)
}
\seealso{
\code{\link{M.rec}},  \code{\link{N.rec}}, \code{\link{I.rec}}, \code{\link{value.rec}}
}
