% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L.rec.R
\name{L.rec}
\alias{L.rec}
\alias{L.rec.matrix}
\title{Times of records}
\usage{
L.rec(X_T, record = "upper")

L.rec.matrix(XM_T, record = "upper")
}
\arguments{
\item{X_T}{A numeric vector.}

\item{record}{A character string indicating the type of record  to be calculated,  "upper" or "lower".}

\item{XM_T}{A numeric matrix.}
}
\value{
A  vector,   containing the times of records.
}
\description{
This function  calculates   the  time (position in the vector)  where  records occur.
}
\details{
The function \code{L.rec.matrix} is equivalent to \code{L.rec} but it is applied to a matrix. The approach to obtain  times of records is applied to each
column of the matrix.  In this function, argument \code{X_T} is  a matrix called \code{XM_T},
and the return value is  a list formed by the  resulting vectors.
}
\examples{
Y1<-c(1,5,3,6,6,9,2)
Y2<-c(10,5,3,6,6,9,2)
Y3<-c(5,7,3,6,19,2,20)
L.rec(Y1)
L.rec.matrix(cbind(Y1,Y2,Y3))
}
\seealso{
\code{\link{I.rec}},\code{\link{M.rec}},  \code{\link{N.rec}}, \code{\link{value.rec}}
}
