% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.test.R
\name{lr.test}
\alias{lr.test}
\title{Likelihood-Ratio Test for the Likelihood of the Record Indicators}
\usage{
lr.test(
  X,
  record = c("upper", "lower"),
  alternative = c("two.sided", "greater", "less"),
  probabilities = c("different", "equal"),
  simulate.p.value = FALSE,
  B = 1000
)
}
\arguments{
\item{X}{A numeric vector, matrix (or data frame).}

\item{record}{A character string indicating the type of record, 
"upper" or "lower".}

\item{alternative}{A character indicating the alternative hypothesis 
(\code{"two.sided"}, \code{"greater"} or \code{"less"}). 
Different statistics are used in the one-sided and two-sided alternatives
(see Details).}

\item{probabilities}{A character indicating if the alternative hypothesis 
assume all series with \code{"equal"} or \code{"different"} probabilities
of record.}

\item{simulate.p.value}{Logical. Indicates whether to compute p-values by
Monte Carlo simulation.}

\item{B}{An integer specifying the number of replicates used in the Monte 
Carlo estimation.}
}
\value{
A list of class \code{"htest"} with the following elements:
  \item{statistic}{Value of the statistic.}
  \item{parameter}{Degrees of freedom of the approximate \eqn{\chi^2} 
    distribution.}
  \item{p.value}{P-value.}
  \item{method}{A character string indicating the type of test.}
  \item{data.name}{A character string giving the name of the data.}
  \item{alternative}{A character string indicating the alternative
    hypothesis.}
}
\description{
This function performs likelihood-ratio tests
  for the likelihood of the record indicators \eqn{I_t} to study the 
  hypothesis of the classical record model.
}
\details{
The null hypothesis of the score tests is that in every vector 
  (columns of the matrix \code{X}), the probability of record at 
  time \eqn{t} is \eqn{1/t} as in the classical record model (i.e., 
  sequences of independent and identically distributed realisations), and 
  the alternative depends on the \code{alternative} and \code{probabilities}
  arguments. The probability at time \eqn{t} is any value, but equal in the
  \eqn{M} series if \code{probabilities = "equal"}  or different in the 
  \eqn{M} series if \code{probabilities = "different"}. The alternative 
  hypothesis is more specific in the first case than in the second one.
  Furthermore, the \code{"two.sided"} \code{alternative} is tested with 
  the usual likelihood ratio statistic, while the one-sided 
  \code{alternatives} use specific statistics based on likelihoods. 
  (See ? (2021) for details on these tests.)

  If \code{alternative = "two.sided" & probabilities = "equal"}, under the
  null, the likelihood ratio statistic has an asymptotic \eqn{\chi^2} 
  distribution with \eqn{T-1} degrees of freedom. It has been seen that for
  the approximation to be adequate \eqn{M} must be between 4 and 5 times 
  greater than \eqn{T}. Otherwise, a \code{simulate.p.value} can be 
  computed.
  
  If \code{alternative = "two.sided" & probabilities = "different"}, the 
  asymptotic behaviour is not fulfilled, but the Monte Carlo approach to 
  simulate the p-value is applied. This statistic is the same as \eqn{\ell} 
  below multiplied by a factor of 2, so the p-value is the same.
  
  If \code{alternative} is one-sided and \code{probabilities = "equal"},
  the statistic of the test is
  \deqn{\mathcal{R} = -2 \sum_{t=2}^T \left\{-S_t \log\left(\frac{tS_t}{M}\right)+(M-S_t)\left( \log\left(1-\frac{1}{t}\right) - \log\left(1-\frac{S_t}{M}\right) I_{\{S_t<M\}} \right) \right\} I_{\{S_t > M/t\}}}
  The p-value of this test is estimated with Monte Carlo simulations,
  since the compute the exact distribution of \eqn{\mathcal{R}} is very 
  expensive.   
  
  If \code{alternative} is one-sided and \code{probabilities = "different"},
  the statistic of the test is
  \deqn{\ell = \sum_{t=2}^T  S_{t} \log(t-1) - M \log\left(1-\frac{1}{t}\right),}
  The p-value of this test is estimated with Monte Carlo simulations. 
  However, it is equivalent to the statistic of the weighted number of 
  records \code{\link{N.test}} with weights \eqn{\omega_t = \log(t-1)}, 
  \eqn{t=2,\ldots,T}.
}
\examples{
set.seed(23)
# two-sided and different probabilities of record, always simulated the p-value
lr.test(ZaragozaSeries, probabilities = "different")
# equal probabilities
lr.test(ZaragozaSeries, probabilities = "equal")
# equal probabilities with simulated p-value
lr.test(ZaragozaSeries, probabilities = "equal", simulate.p.value = TRUE)

# one-sided and different probabilities of record
lr.test(ZaragozaSeries, alternative = "greater", probabilities = "different")
# different probabilities with simulated p-value
lr.test(ZaragozaSeries, alternative = "greater", probabilities = "different", 
  simulate.p.value = TRUE)
# equal probabilities, always simulated the p-value
lr.test(ZaragozaSeries, alternative = "greater", probabilities = "equal")
}
\references{
? (2021).
“Statistical Tests to Detect Non-Stationarity Based on Records to Analyse Climate Change.”
Unpublished manuscript.
}
\seealso{
\code{\link{score.test}}, \code{\link{global.test}}
}
\author{
Jorge Castillo-Mateo
}
