% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_rev.R
\name{series_rev}
\alias{series_rev}
\title{Reverse elements by columns}
\usage{
series_rev(XM_T)
}
\arguments{
\item{XM_T}{A numeric matrix (or data frame).}
}
\description{
Result from apply \code{rev} function by columns to the matrix.
  This allows studying the series backwards and not just forward.
}
\examples{
series_rev(matrix(1:100, 10, 10))

series_rev(ZaragozaSeries)

}
\seealso{
\code{\link{series_double}}, \code{\link{series_split}}, 
  \code{\link{series_uncor}}, \code{\link{series_untie}}
}
\author{
Jorge Castillo-Mateo
}
