% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TX_Zaragoza-data.R
\docType{data}
\name{TX_Zaragoza}
\alias{TX_Zaragoza}
\title{Time Series of Daily Maximum Temperature at Zaragoza (Spain)}
\format{
A data frame with 28670 rows and 5 variables.
}
\source{
\href{https://www.ecad.eu}{EUROPEAN CLIMATE ASSESSMENT & DATASET (ECA&D)}
}
\usage{
data(TX_Zaragoza)
}
\description{
A dataset containing the series of daily maximum temperature at 
  Zaragoza (Spain), from 01/01/1953 to 31/12/2018. This series is obtaned 
  from the ECA series but it has been transformed, by removing days February 
  29th and imputing the two missing values. The variables are the following:

\itemize{
  \item STAID : Station identifier
  \item SOUID : Source identifier
  \item DATE  : Date YYYYMMDD
  \item TX    : Maximum temperature in 0.1 ºC
  \item Q_TX  : quality code for TX (0='valid'; 1='suspect'; 9='missing')
}
}
\references{
Klein Tank AMG and Coauthors (2002). 
Daily Dataset of 20th-century Surface Air Temperature and Precipitation 
Series for the European Climate Assessment.
\emph{International Journal of Climatology}, \strong{22}(12), 1441-1453. 
\href{https://doi.org/10.1002/joc.773}{doi:10.1002/joc.773}
}
\seealso{
\code{\link{ZaragozaSeries}}
}
\keyword{datasets}
