% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZaragozaSeries-data.R
\docType{data}
\name{ZaragozaSeries}
\alias{ZaragozaSeries}
\title{ZaragozaSeries}
\format{A matrix with 33 rows and 150 columns.}
\usage{
data(ZaragozaSeries)
}
\description{
The  matrix resulting from the preprocessing  of  TX_Zaragoza$TX.  The matrix is the result from applying  consecutively:
 \code{\link{DaySeries.fun}}, \code{\link{mintime.cor.test}} and \code{\link{double.fun}}.
}
\keyword{datasets}
