% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_rev.R
\name{series_rev}
\alias{series_rev}
\title{Reverse Elements by Columns}
\usage{
series_rev(X)
}
\arguments{
\item{X}{A numeric vector, matrix (or data frame).}
}
\description{
Result of applying \code{\link[base]{rev}} function by columns
  to the matrix. This allows the study of the series backwards and not only 
  forward.
}
\examples{
series_rev(matrix(1:100, 10, 10))

series_rev(ZaragozaSeries)

}
\seealso{
\code{\link{series_double}}, \code{\link{series_record}}, 
  \code{\link{series_split}}, \code{\link{series_ties}},
  \code{\link{series_uncor}}, \code{\link{series_untie}}
}
\author{
Jorge Castillo-Mateo
}
