% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZaragozaSeries-data.R
\docType{data}
\name{ZaragozaSeries}
\alias{ZaragozaSeries}
\title{Splitted and Uncorrelated Time Series \code{\link{TX_Zaragoza}}}
\format{
A matrix with 66 rows and 76 columns.
}
\usage{
data(ZaragozaSeries)
}
\description{
The matrix resulting from the pre-processing of 
  \code{\link{TX_Zaragoza}$TX}.
}
\details{
The matrix is the result from applying:
  \code{\link{series_uncor}(\link{series_split}(\link{TX_Zaragoza}$TX))}.
}
\seealso{
\code{\link{TX_Zaragoza}}
}
\keyword{datasets}
