% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZaragozaSeries-data.R
\docType{data}
\name{ZaragozaSeries}
\alias{ZaragozaSeries}
\title{Splitted and Uncorrelated Time Series \code{\link{TX_Zaragoza}}}
\format{
A matrix with 70 rows and 76 columns.
}
\usage{
data("ZaragozaSeries")
}
\description{
The matrix resulting from the data preparation (or 
  pre-processing) of \code{\link{TX_Zaragoza}$TX}.
}
\details{
The matrix is the result from applying:
  \code{\link{series_uncor}(\link{series_split}(\link{TX_Zaragoza}$TX))}.
  
  The data matrix corresponds to the 70 years with observations in 
  \code{\link{TX_Zaragoza}$TX} and to the 76 days in the year where adjacent
  daily maximum temperature subseries are uncorrelated. By coincidence, 
  none of the subseries 4, 90 or 278 with missing observations is kept 
  within the 76 uncorrelated days.
}
\seealso{
\code{\link{TX_Zaragoza}}
}
\keyword{datasets}
