% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZaragozaSeries-data.R
\docType{data}
\name{ZaragozaSeries}
\alias{ZaragozaSeries}
\title{Splitted and uncorrelated time series \code{\link{TX_Zaragoza}}}
\format{
A matrix with 66 rows and 75 columns.
}
\usage{
data(ZaragozaSeries)
}
\description{
The matrix resulting from the preprocessing of 
  \code{TX_Zaragoza$TX}. The matrix is the result from applying consecutively:
  \code{\link{series_split}} and \code{\link{series_uncor}}.
}
\keyword{datasets}
